update stamping of non image or imu messages
This script current copies all messages into a new bag based on header time, instead of index time.
This is fine for messages that we are restamping, because that'll make them playback in more 'real-time'
However, we can't do this for all message types, because some message types don't actually have a header.
I ran into this problem with the vicon bags, because they have some messages that don't contain a header so this script crashes. This fix solved that.